Add files via upload
Task 1:

1.1
Write a Python Program to implement your own myreduce() function which works exactly like
Python's built-in function reduce()
1.2
Write a Python program to implement your own myfilter() function which works exactly like
Python's built-in function filter()
2.
Implement List comprehensions to produce the following lists.
Write List comprehensions to produce the following Lists

['x', 'xx', 'xxx', 'xxxx', 'y', 'yy', 'yyy', 'yyyy', 'z', 'zz', 'zzz', 'zzzz']
['x', 'y', 'z', 'xx', 'yy', 'zz', 'xx', 'yy', 'zz', 'xxxx', 'yyyy', 'zzzz']
[[2], [3], [4], [3], [4], [5], [4], [5], [6]]
[[2, 3, 4, 5], [3, 4, 5, 6], [4, 5, 6, 7], [5, 6, 7, 8]]
[(1, 1), (2, 1), (3, 1), (1, 2), (2, 2), (3, 2), (1, 3), (2, 3), (3, 3)]
3.
Implement a function longestWord() that takes a list of words and returns the longest one.
Task 2:

1.1
Write a Python Program(with class concepts) to find the area of the triangle using the below
formula.
area = (s*(s-a)*(s-b)*(s-c)) ** 0.5
Function to take the length of the sides of triangle from user should be defined in the parent
class and function to calculate the area should be defined in subclass.
1.2
Write a function filter_long_words() that takes a list of words and an integer n and returns the list
of words that are longer than n.
2.1
Write a Python program using function concept that maps list of words into a list of integers
representing the lengths of the corresponding words.
Hint: •If a list [ ab,cde,erty] is passed on to the python function output should come as [2,3,4]

Here 2,3 and 4 are the lengths of the words in the list.
2.2
Write a Python function which takes a character (i.e. a string of length 1) and returns True if it is
a vowel, False otherwise.